pyproject.toml
[project]
name = "files-manager-module"
version = "0.1.0"
description = ""
authors = [
    {name = "Khululu",email = "alonzost1987@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow (>=11.1.0,<12.0.0)",
    "pillow-heif (>=0.21.0,<0.22.0)",
]

[tool.poetry]
packages = [{include = "files_manager_module", from = "."}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
